requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.28
Werkzeug==3.0.1
gunicorn==21.2.0
python-dotenv==1.0.1
Jinja2==3.1.3
itsdangerous==2.1.2
click==8.1.7 